fix: api 오류 처리 개선 및 api 주소 변경
